Change structure of project (create separate files for classes), use scoped_lock in method use(), add method getMutex()
