+++ PraceDomowe/.gitignore
@@ -1,1 +1,2 @@
 /a.out
+/zad4_dining_philosophers/a.out
