+++ PraceDomowe/zad4_dining_philosophers/Fork.cpp
@@ -1,0 +1,6 @@
+#include "Fork.hpp"
+
+std::mutex& Fork::getMutex()
+{
+    return forkMutex;
+}
