+++ PraceDomowe/zad4_dining_philosophers/Fork.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <mutex>
+
+class Fork{
+private:
+    std::mutex forkMutex;
+
+public:
+    std::mutex& getMutex();
+};
