+++ PraceDomowe/zad4_dining_philosophers/Philosopher.hpp
@@ -1,0 +1,23 @@
+#pragma once
+
+#include <iostream>
+#include <mutex>
+#include <string>
+#include <thread>
+#include "Fork.hpp"
+
+class Philosopher{
+private:
+    std::string philosopherName;
+    Fork& leftFork;
+    Fork& rightFork;
+    std::thread philosopherThread;
+
+public:
+    Philosopher(std::string name, Fork lFork, Fork rFork);
+    ~Philosopher();
+
+    void eat();
+    void think();
+    void dine();
+};
