+++ PraceDomowe/zad4_dining_philosophers/zad4_dining_philosophers.cpp
@@ -1,0 +1,7 @@
+#include "Fork.hpp"
+#include "Philosopher.hpp"
+
+int main(){
+
+    return 0;
+}
